misc: move `HandlerFunc` into a `testing` package
This moves the `HandlerFunc` into a `testing` package to avoid returning it as part of our public API.
